chore:created new helper function file
