+++ admin/dashboard/assets/src/dashboard-app/pages/fonts/GoogleFont.js
@@ -1,58 +1,17 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { useDispatch } from 'react-redux';
-import { __ } from "@wordpress/i18n";
+import { __ } from '@wordpress/i18n';
 import { useSelector } from 'react-redux';
 import apiFetch from '@wordpress/api-fetch';
+import getFontWeightTitle from '../../../common/fontcommon';
 
 const GoogleVariationItem = ({
 	id,
 	variation,
 	handleVariationRemove
 }) => {
-	const getFontWeightTitle = ( weight ) => {
-		if ( undefined === weight ) {
-			weight = '400';
-		}
-		let updatedWeight = weight,
-			oldWeight = weight;
-		if ( 'italic' === weight ) {
-			oldWeight = '400italic';
-		}
-		if ( oldWeight.includes('italic') ) {
-			updatedWeight = `${oldWeight.replace('italic', '')} Italic`;
-		}
-		switch ( weight ) {
-			case '100':
-			case '100italic':
-				return __( 'Thin ', 'custom-fonts' ) + updatedWeight;
-			case '200':
-			case '200italic':
-				return __( 'Extra Light ', 'custom-fonts' ) + updatedWeight;
-			case '300':
-			case '300italic':
-				return __( 'Light ', 'custom-fonts' ) + updatedWeight;
-			case '400':
-			case '400italic':
-				return __( 'Regular ', 'custom-fonts' ) + updatedWeight;
-			case '500':
-			case '500italic':
-				return __( 'Medium ', 'custom-fonts' ) + updatedWeight;
-			case '600':
-			case '600italic':
-				return __( 'Semi Bold ', 'custom-fonts' ) + updatedWeight;
-			case '700':
-			case '700italic':
-				return __( 'Bold ', 'custom-fonts' ) + updatedWeight;
-			case '800':
-			case '800italic':
-				return __( 'Extra Bold ', 'custom-fonts' ) + updatedWeight;
-			case '900':
-			case '900italic':
-				return __( 'Ultra-Bold ', 'custom-fonts' ) + updatedWeight;
-			default:
-				return updatedWeight;
-		}
-	}
+	useEffect(()=>{getFontWeightTitle(weight);
+	},[])
 
 	return (
 		<div key={id} className="flex items-center justify-between">
